requirements.txt
langchain
langchain_community
langchain-google-genai
python-dotenv
streamlit
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
pypdf
rapidocr-onnxruntime
fastapi
uvicorn[standard]
SQLAlchemy
python-multipart
requests
python-docx
PyJWT
passlib[bcrypt]
jinja2
chromadb
langchain-text-splitters
